requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
jsonschema==4.21.1